requirements.txt

openai>=1.0.0
rich>=13.0.0
prompt-toolkit>=3.0.0
requests>=2.31.0
PyYAML>=6.0.0
tiktoken